Remove XSCT path check
Issue #17 presents an error that occurs when only Vivado and not Vitis is installed. The check that was forcing this error was redundant anyway, as the xsct binary included in the Vitis install path is never used by the scripts, and was a holdover from when some Xilinx SDK suport was provided.